build: add missing dependency on MLIRTorchTypesIncGen
Because of the missing dependency, CI saw non-deterministic failures,
where the build failed because the TorchTypes.h.inc file was not
available to include, likely due to the order in which the targets
happened to be compiled.  This patch fixes the problem by making sure
that the TorchTypes.h.inc file is generated before it is referenced.