build: pin most recent nightly PyTorch release
Instead of using the git commit hash and then building PyTorch from
source, this patch fetches the most recent nightly (binary) build of
PyTorch, before pinning it in pytorch-requirements.txt, which is
referenced in the top-level requirements.txt file.  This way, end users
will continue to be able to run `pip -r requirements.txt` without
worrying whether doing so will break their Torch-MLIR build.

This patch also fetches the git commit hash that corresponds to the
nightly release, and this hash is passed to the out-of-tree build so
that it can build PyTorch from source.