build: sort versions instead of numbers
If we were to sort the input as numbers (in descending order), then 1.9
appears before 1.13.  To fix this problem, we use the `--version-sort`
flag (along with `--reverse` for specifying a descending order).  We
also filter out lines that don't contain version numbers by only
considering lines that start with a digit.