build: use Github Actions for creating PR
Prior to this patch, the CI script used the Github CLI (`gh`) to create
a PR, but doing so had the disadvantage that the created PR was
attributed to Sean.  This patch uses the
`peter-evans/create-pull-request@v4` Github action which allows us to
customize various parts of the PR, including the committer's name.  By
using a specific commiter name, we can also set the PyTorch version
update PRs to be auto-approved.