build: retain object files after building PyTorch from source
With the goal of improving time spent in CI, this patch retains the
object files built while compiling PyTorch from source.  By doing so,
the build system would recognize that the object files are newer than
the source files, thus skipping a full rebuild.

Although this patch does not make any _functional_ changes, it still is
speculative.  So if we do not observe any significant improvement in
time spent in CI over the next few days, I will revert this patch so
that we can save some disk space.