Move PrimLoop, ListOp, CastOp patterns to abstract interp utils
This commit moves the `FullyUnrollPrimLoopOp`,
`FoldPrimUncheckedCastOp` and `AbstractlyInterpretListOpsWithinABlock`
patterns to the `SimplifyAbstractInterpCalculationsUtils.cpp` file so
that the dtype simplification pipeline can also use them.

In addition, the `Aten__Getitem__TOp` and `PrimTupleUnpackOp`
canonicalizers are added to the dtype simplification pipeline to deal
with the ops generated by the `aten.cat` dtype function.