Add dtype functions for rand* and var_mean* ops (#2089)
* Add dtype functions for rand* ops

This commit adds the dtype functions for the rand* ops. It also makes
a small change to the `check_dtype_function` implementation to use the
underlying int representation of the dtype when comparing the result
dtype and the golden dtype. This is needed because one of the dtype
functions has a default dtype parameter, which is represented as an
int in the registry.

* Add dtype functions for var_mean* ops